requirements.txt
streamlit
python-dotenv
google-genai
torch
transformers
tensorflow
keras
joblib
pandas
numpy
opencv-python
gtts
soundfile
openai-whisper
pydub
ffmpeg-python
